Changed directory name to tests
